+++ next/src/app/page.tsx
@@ -1,4 +1,4 @@
-import { fetchSanity, groq } from '@/lib/sanity'
+import { creativeModuleQuery, fetchSanity, groq } from '@/lib/sanity'
 import Modules from '@/ui/modules'
 import processMetadata from '@/lib/processMetadata'
 
@@ -25,16 +25,7 @@ async function getPage() {
 						internal->{ title, metadata }
 					}
 				},
-				modules[]{
-					...,
-					ctas[]{
-						...,
-						link{
-							...,
-							internal->{ title, metadata }
-						}
-					}
-				}
+				${creativeModuleQuery}
 			},
 			metadata {
 				...,